Organised assemblies and namespaces to match the Duck template repo
