--- Scripts/HierarchyBehaviour.asmdef
@@ -1,3 +1,0 @@
-﻿{
-	"name": "HierarchyBehaviour"
-}
