+++ Tests/Behaviours/Duck.HierarchyBehaviour.Tests.Behaviours.asmdef
@@ -1,0 +1,10 @@
+{
+    "name": "Duck.HierarchyBehaviour.Tests.Behaviours",
+    "references": [
+        "Duck.HierarchyBehaviour"
+    ],
+    "optionalUnityReferences": [],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false
+}
